move helper data structures to utils directory
