+++ src/token_array.hpp
@@ -1,5 +1,5 @@
 #pragma once
 
-#include "chunked_array.hpp"
+#include "utils/chunked_array.hpp"
 
 using TokenArray = ChunkedArray<Token>;
